Prevent failed 'finalize_job_results' Minion jobs for skipped jobs
I'm not sure why openQA jobs are sometimes immediately cancelled (after
just being scheduled). This change prevents enqueuing a
'finalize_job_results' Minion job for these openQA jobs because they
sometimes fail (see https://progress.opensuse.org/issues/75073). This can
also be seen as an optimization because the Minion jobs are not required
anyways.

The enqueuing has been moved out of store_column because this function is
invoked by DBIx for each column individually in a random order. Since the
condition for enqueuing the job depends on both - state and result - it can
not (easily) be evaluated within store_column. It is likely cleaner to
enqueue the job explicitly within done/cancel which are the only places
where it actually makes sense.

There's still the question why we sometimes see these immediately cancelled
jobs in the first place. They all have in common to be part of a parallel
cluster which has been duplicated for some reason. (Some of the jobs in the
parallel cluster also have a chained parent.)