docker: Ask to not squash base image layers to fix circleCI download
Reference:
https://openbuildservice.org/help/manuals/obs-user-guide/cha.obs.building.html#id-1.5.8.3.8

The combination of a heavy image in combination with downloading from
slow registry.opensuse.org to circleCI workers have a significant impact
causing frequent failures in the "cache" step of circleCI when the 542MB
blob of the base image is downloaded. As we can not configure circleCI
to retry the download or apply a longer timeout than 20 minutes this
means that manual interaction is necessary to retry.

This commit changes the base image to preserve individual layers within the
container image which makes downloading much more likely to succeed.

Related progress issue: https://progress.opensuse.org/issues/67855