circleci: Strip down base image to prevent "cache" jobs failing
The combination of a heavy image in combination with downloading from
slow registry.opensuse.org to circleCI workers have a significant impact
causing frequent failures in the "cache" step of circleCI when the 542MB
blob of the base image is downloaded. As we can not configure circleCI
to retry the download or apply a longer timeout than 20 minutes this
means that manual interaction is necessary to retry.

This commit strips down the base image to the bare minimum to only
provide what circleCI needs to save/restore the cache and changes the
circleCI configuration to use what a vanilla openSUSE Leap container
provides, which is also much smaller. At time of writing 42MB. Then we
rely on the package install within the "cache" step to install all
dependencies for which we can also apply timeouts and retry values
ourselves.

Related progress issue: https://progress.opensuse.org/issues/67855