Handle placeholders after parsing _URL settings
b3bc8ebf caused some problems by moving the parsing of `_URL`
settings to after the call to `JobSettings::generate_settings`,
which is where handling of placeholders (%FOO%), usually from the
templates, is done. This broke the case where e.g. a template
uses `%ISO%` in a value, and the job is posted with `ISO_URL` but
no explicit `ISO` set. In that case the derived `ISO` setting
should be used in the placeholder expansion, but this was broken.

This fixes the problems by splitting up the "parse _URL settings"
and "create download list" functions of `create_downloads_list`,
and moving the parsing part into `generate_settings`, just before
placeholder expansion happens. This is after we pull in settings
from the templates, so the bug b3bc8ebf was trying to fix (POO
62159) should still be fixed. Download list creation should still
happen the same as currently, so POO #70687 should still be fixed.

Signed-off-by: Adam Williamson <awilliam@redhat.com>