Re-do handling of placeholders and + after _URL variables parsed
b3bc8ebf caused some problems by moving the parsing of `_URL`
settings to after the call to `JobSettings::generate_settings`,
which is where handling of placeholders (%FOO%) and +-overrides
(+FOO), usually from the templates, is done. This broke the case
where e.g. a template sets `+ISO`, or uses `%ISO%` in a value,
and the job is posted with `ISO_URL` but no explicit `ISO` set.
In that case the derived `ISO` setting should be overridden or
used in the placeholder expansion, but this was broken.

This fixes the problems by handling placeholder expansion and
+-overrides *again* after the parsing of `_URL` settings. They
are pretty simple operations anyway so it shouldn't cause much
overhead. To facilitate this we have to set up destructive and
non-destructive modes for `handle_plus_in_settings` so it can be
run non-destructively at first and then destructively later, and
similarly set up definitive and non-definitive modes for
`expand_placeholders`, so on the first run if the setting a
placeholder refers to is not (yet?) set, we leave the placeholder
rather than replacing it with an empty string.

Signed-off-by: Adam Williamson <awilliam@redhat.com>