Fix UI build bar alignment broken on index page
Problem: The right side of some build bars of is
misaligned when the screen is less wide than 880
pixels. This occurs because some bars have more
text than others and this text forces a minimum size.

Solution: a minimum size is set for all bars to
ensure the maximum possible text size.

https://progress.opensuse.org/issues/77929