Don't set FOO from FOO_URL if FOO is defined but false
If someone has explicitly set e.g. ISO to "" for some scenario,
that's a clear signal they want the test in that scenario to run
without an ISO attached, even if it's posted with ISO_URL set.
However, when we decide whether to generate FOO from FOO_URL,
the check we use is not whether FOO is already *defined* in the
settings hash, but whether it *evaluates true*. That seems wrong.

We ran into this is in real life after b3bc8eb changed the order
in which we do _URL parsing and +-overrides. There is a Fedora
test suite with `+ISO` set to `''`, the intent being that the
'ISO' value derived from the posted `ISO_URL` should be
overridden to `''` when running that test (so it runs with no
disc attached, and boots from the network). This worked so long
as we handled the +-overrides after we parsed the _URL settings,
but b3bc8eb changed things so we do the +-overrides before we
parse the _URL settings, and caused that case to trigger this.

The previous commit to this one changes the ordering back to fix
a similar issue with placeholder expansion, and so also happens
to hide this bug again in that specific real life case, but this
seems like the 'correct' fix. Consider for instance if the test
suite just set `ISO` (not `+ISO`) to `''`; it seems like even
then we ought not to override that intent by parsing `ISO_URL`,
it should not actually be necessary for the test suite to use
the +-override mechanism.

Signed-off-by: Adam Williamson <awilliam@redhat.com>