parse_url_settings: improve filename parse check
There's no need to set the value then delete it if it turns out
we set it to an empty string; that's a hangover from when we did
not assign `$filename` but set the value directly to the return
of the `Mojo::URL` call. Let's just check if we managed to get a
filename and bail before setting the value if not. This also
reduces potential confusion around truthiness vs. defined checks.

Signed-off-by: Adam Williamson <awilliam@redhat.com>