parse_url_settings: don't parse value for non-asset types
If a test sets NOT_ASSET_URL: http://some/parseable.url , we don't
actually want to parse that to NOT_ASSET: parseable.url , because
we aren't going to download it later. We had this bug back in 2016
and fixed it specifically in f54a18f , then regressed it again in
391f95e3 because we needed to pass the parsed file name to the
asset type check.

When we split up `parse_url_settings` and `create_downloads_list`
the asset type check wound up in `create_downloads_list`. To fix
the bug we can't avoid duplicating it here - we do have to check
it in both places, here to avoid setting $settings->{$short} if
it's not an asset type, and there to avoid doing the download if
somehow both NOT_ASSET_URL and NOT_ASSET wind up being set by
templates or the scheduler. It's not worth splitting the check to
a shared utility function as you still need flow control. I don't
think the log message will ever be duplicated as we should never
manage to reach and fail the check *both* times (only one or the
other).

Signed-off-by: Adam Williamson <awilliam@redhat.com>