systemd: Prevent error reports in journal of openqa-auto-update service
The service openqa-auto-update might trigger an update of the package
that includes the service openqa-auto-update itself which would trigger
a stop request for the service which can not be fulfilled as rpm does
not react on termination when trying to stop a systemd service causing a
deadlock until the service is killed which triggers errors despite
everything else seems to work just fine.

This commit changes the package pre/post scripts to only touch the timer
and not the service. I think it's fine because the script that is
started by the service has nothing to do with any other files from the
package. So I consider it safe because the script would already be fully
in memory during execution while the according service and package might
be replaced in an upgrade.

This is inspired by
https://src.fedoraproject.org/rpms/dnf/blob/master/f/dnf.spec
following the same approach.

Thanks to "defolos" for the hints and support while trying to dig into
this.

Related progress issue: https://progress.opensuse.org/issues/70978