Add 'INIT_PRODUCTDIR' to record the defult needle_dir
When we use `openqa-clone-custom-git-refspec` to trigger a job that is
run on a remote worker, the new job may fail because it cannot find the
needle dir. This is because not all workers' cache service is enabled and
the WebUI host is defined differently among these workers.

See more details in: https://progress.opensuse.org/issues/67723