t: Fix local execution of 16-utils-runcmd depending on filesystem layout
git distinguishes between
"not a git repository (or any of the parent directories)"
and
"not a git repository (or any parent up to mount point %s)"
which can lead to different error messages that our tests may show
depending on the local development environment encountered. See
https://github.com/git/git/blob/7814e8a05a59c0cf5fb186661d1551c75d1299b5/setup.c#L1245
and
https://github.com/git/git/blob/7814e8a05a59c0cf5fb186661d1551c75d1299b5/setup.c#L1251
accordingly.

This commit generalizes the expression to not fail in an environment
where previously the message "not a git repository (or any parent up to
mount point %s)" was encountered. Also reducing the line breaks within
statements.