Fix web UI container script in basic execution
When web UI is executed in a container executes the script run_openqa.sh
if this is executed in the default mode with all components together
uses the function all_together_apache(). This function has some errors
that avoid the execution of the scheduler, websockets and livehandler
because some parameters for the command start_daemon are missing.

Additionally the same simplification applied here to the gru daemon
is also applied to the function gru() which is used by the load balance
execution mode.

https://progress.opensuse.org/issues/80534