Add container test for openQA worker
The test probes that the container image can be build and can be used
deploying a worker. But because the are not an appropriate environment
is going to fail in the execution. The test checks that show some part
of this information

The current execution with an appropriate environment and configuration
is

----
[info] [pid:20] worker 1:
 - config file:           not found
 - worker hostname:       ce39563038b7
 - isotovideo version:    20
 - websocket API version: 1
 - web UI hosts:          localhost
 - class:                 ?
 - no cleanup:            no
 - pool directory:        /var/lib/openqa/pool/1
[error] [pid:20] Errors occurred when reading config file:
 - Config file not found at '/etc/openqa/workers.ini'.
API key and secret are needed for the worker connecting localhost
----

https://progress.opensuse.org/issues/80682