docker: Include the current web UI build alternatives
Currently the web UI can be compiled in three different ways,
to be used with all its services in the same container using
apache as a reverse proxy with or without SSL certificate,
or with its services separately to be used with nginx as a
reverse proxy for example in the docker-compose case.

This PR provides a docker file for each of the different models.

https://progress.opensuse.org/issues/80520