docker: Add a test for openQA workers containers
This PR add a test for the openQA workers containers

The current execution with an appropriate environment and configuration
is

----
[info] [pid:21] worker 1:
 - config file:           /etc/openqa/workers.ini
 - worker hostname:       f2fa029ca2ae
 - isotovideo version:    20
 - websocket API version: 1
 - web UI hosts:          http://localhost:9526
 - class:                 qemu_i386,qemu_x86_64
 - no cleanup:            no
 - pool directory:        /var/lib/openqa/pool/1
API key and secret are needed for the worker connecting
http://localhost:9526
----

The test probes that the container image can be build and can be used
deploying a worker. But because the is not a web UI is going to fail
in the execution. The test checks that show some part of this information

progress.opensuse.org/issues/80682