Avoid worker from appearing stuck (e.g. not reacting to SIGTERM) on cache check
When the worker checks for the availability of the cache service the normal
behavior of the cache service client is applied (huge number of attempts,
high inactivity timeout). This is not nice for the availability check
because it leads to the worker getting stuck and not even reacting to
signals.