Mock the Minion framework in tests for df-based cleanup to speed up coverage analysis
This speeds up the test execution significantly when coverage analysis is
enabled. Locally it now takes only a few seconds anymore. Before it timed
out after 30 seconds. In the CI (which has an extended timeout) it timed
out after 60 seconds.

The same approach could be applied to other affected test as well
because the mocking is a drop-in replacement for the existing test helper.