Don't rely on re-checking statfs() during the cleanup
In my tests it seems that statfs() might not immediately reflect the a
deletion causing the cleanup to delete too many jobs. Hence it is better
not to rely on statfs() during the cleanup and only use it for the initial
check. This might impair the performance as we need to track the size of
deleted results and screenshots now manually but it is better than deleting
too many jobs.