script: Fallback to running openqa-bootstrap-container without pipe
Older versions of systemd-run do not have the `-P` parameter. We
originally envisioned that newer versions of Leap would get support but
as the main version of systemd stays the same as in the SLE15 codestream
understandably also the parameters stay the same.

This commit actively checks if the parameter is available. If it is
available it is used, otherwise it will output a message to the user
accordingly.

This is based on a user support question brought up in text chat.