Fix problem in fullstack test leading to occasional failures
Sometimes the actual error message for the failed download is available and
therefore that message is logged (instead of the more generic one). The
reason should always contain the "Failed to download…" message so I'm
adding a check for that.

So far I've noticed that problem only when testing locally (to reproduce
https://progress.opensuse.org/issues/80800 which is **not** fixed by this
change).