Set an absolute path as PRODUCTDIR
1. In the last version, when users set a github address as CASEDIR and don't
define the PRODUCTDIR, the PRDUCTDIR will be a relative path. This will
cause a job cannot find 'main.pm'. In this scenario, set the absolute
path as PRODUCTDIR.
2. When users use `openqa-clone-custom-git-refspec` to trigger a job
whose CASEDIR is a basename (such as opensuse), the NEEDLES_DIR will be
set `opensuse/products/opensuse/needles`, but this needles_dir does not
exist. In this scenario, set the basename `needles` as `NEEDLES_DIR`,
and do symlink.