Allow to override isotovideo executable with command line
os-autoinst includes multiple backends which can require changes to
support different technologies or specific testing needs. Also, the
worker environment needs to provide requirements to run tests which
include more package than originally required for os-autoinst or newer
versions.

This commit introduces the possibility to customize the actual worker
engine that is started with keeping the default of starting "perl
$isotovideo -d". For example a different location for isotovideo can be
specified or a container providing different versions of packages or
even a completely different worker engine that is interface-compatible.

https://progress.opensuse.org/issues/90362