Treat result uploads for non-existant modules as error
The worker reads the test order and updates it before each upload attempt
so the web UI should be aware of all test modules even if the test schedule
changes in the middle. So it should be safe to treat this situation as an
error.

This should improve the error feedback in cases like
https://progress.opensuse.org/issues/90152 where the logs tell us currently
that the result upload worked but in the end nothing is there. Of course
this just improves the symptom. The cause is likely that the test order
could not be uploaded in the first place.