Fix several more API routes for Mojolicious 9.11 routing change
Route format detection is disabled by default from Mojolicious
9.11. This breaks accessing, for instance, /overview.json and
expecting it to be handled the same as accessing /overview with
a JSON content-type header. We have several routes that are
implemented to allow for this. #1820 adjusted several routes
where the tests actually use format detection to make it work
again; this adjusts several remaining routes that also would
previously have worked this way, but where no test happened to
use it. Without this change, those routes are accessible with
format detection with Mojo <9.11, but not with Mojo >=9.11.
With this change, they work with both, at the minor 'cost' that
URLs generated for them with `url_for` will now have a .html
extension where they did not before.

Signed-off-by: Adam Williamson <awilliam@redhat.com>