containers: Add healthchecks to organize the docker-compose startup
Eventually the startup using docker-compose fails several seconds
after all is up. The reason is because the webui do operations over
the DB and because we have concurrent webui containers can cause
race conditions

For instance:

Webui_1 failed to run SQL in
/usr/share/openqa/script/../dbicdh/PostgreSQL/deploy/92/001-auto-__VERSION.sql:
DBIx::Class::DeploymentHandler::DeployMethod::SQL::Translator::try {...} ():
 DBI Exception: DBD::Pg::db do failed:
ERROR:  duplicate key value violates unique constraint "pg_type_typname_nsp_index"

An example of this failure could be found at:
https://github.com/os-autoinst/openQA/pull/3818

The reason for the problem is that all webui containers try to create
and initialize the database at the same time, and since all webui
replicas are launched at the same time by the docker-compose and we
have more than one, they overlap and generate the issue

The solution approach has been to create an initial webui container with
a health check. This is the first container in the entire web UI system.
So from it, using a cascade of appropriate dependencies so that the
whole system starts up in an orderly way. And the health checks allow
that the next step is not going to start until the previous one is
not complete.

progress.opensuse.org/issues/89731