Ensure workers are shown as offline after disconnecting gracefully immediately
The code so far relied on the fact that `t_seen` is unset on a graceful
disconnect. However, in practice that does not always work, likely due to
some race condition. So let's just not rely on that. Considering any worker
offline with the error 'gracefully disconnected at …' is simple and not a
problem because that error is unset when the worker connects again.