CI: Fix 01-test-utilities.t
It can happen that the warning in the subprocess happens before the
SIGCHLD handler was installed.

The sleep will fix this in most cases.

(To be really sure one would have to add some method in Utils.pm to check if
the handler has been installed.)

Also I changed combined_like to combined_from, which made it easier to debug
the actual output.

Issue: https://progress.opensuse.org/issues/92164