Tweak routing again to work for Mojo < 9 or >=9.11
We cut off support for Mojo before 9 pretty suddenly with #3830,
and it leaves Fedora in an awkward position as F34 shipped with
Mojo 8 and updating to 9 within a stable release would be against
Fedora's update policies. This adjusts the routing to work the
old way (before #3816, #3820 or #3830) for Mojolicious versions
before 9, and the new way for versions newer than 9.10; for
anything in the middle, we die with an error.

Signed-off-by: Adam Williamson <awilliam@redhat.com>