worker: Fix handling CASEDIR/PRODUCTDIR/NEEDLES_DIR in certain cases
This change ensures that the default-assignment for PRODUCTDIR is a
relative directory (unless the feature switch to use absolute paths is
turned on). That is required when a custom CASEDIR has been specified
because in this case it is expected that not only Perl modules but also
`main.pm` itself is loaded from that custom CASEDIR (unless PRODUCTDIR is
set explicitly to some other location). So this should fix
https://progress.opensuse.org/issues/93101.

When only changing the behavior for PRODUCTDIR the needles could not be
loaded anymore (unless NEEDLES_DIR is specified explicitly) because the
checkout of CASEDIR does not necessarily contain needles. To fix this, the
symlinking of the needles directory is done in that case. See the verbose
comments within the code for further explanations.

Note that the absolute PRODUCTDIR has been introduced by the commit
55420c174299d545d654d4c6266415a8e9ee8dd2 to fix two problems. The first
problem won't occur again because now the relative path for PRODUCTDIR is
computed correctly. The second problem is the needle problem mentioned in
the previous paragraph and thus handled as well.

Tested locally with
* no custom CASEDIR, NEEDLES_DIR and PRODUCTDIR
* custom CASEDIR, NEEDLES_DIR, PRODUCTDIR as produced by assigned by
  `openqa-clone-custom-git-refspec`
* only a custom CASEDIR as assigned by `openqa-investigate` (case mentioned
  in https://progress.opensuse.org/issues/93101)
* only a custom NEEDLES_DIR pointing a some Git repository