Fix problem in t/ui/25-developer_mode.t leading to sporadic test failures
The mocking of status updates was insufficient. It is not enough to
override `updateStatus()` after loading the page because it might have
already been invoked and therefore the initial status update is not
actually prevented. This change allows to disable status updates from the
beginning via a query parameter (which is only evaluated in test mode).