Delete uncovered and likely unnecessary code for "failed_modules" parsing
"failed_modules" is supported as a query parameter on /tests/overview
and works with the existing code in the API controller sorting out
modules by result. The additional search parameter on the database is
neither necessary nor called anywhere in our tests. "complex_query" is
used by /tests as well but there the query parameter "failed_modules"
already does not work hence we can remove the dead code in the backend.
The code in the HTML form was removed in 4865792 but forgotten in the
backend.