Optimize search: prefetch job columns
Currently for every job-module result a `SELECT FROM job` is executed,
when calling $job_module->job->name.

The prefetch will save the job data in the object and so the extra
SELECT is avoided.
Additionally this commit reduces the selected columns to what is actually
needed.

A search for 'vlc' was reduced from 39s to 7s locally.