Ensure we end testing when bailing due to missing Selenium drivers
There was a problem in `27-plugin_obs_rsync_status_details.t`; because
`setup_obs_rsync_test` does a few things that are identified as 'tests',
if we run it then want to bail out because drivers are missing, we need
to end testing somehow. Thanks to @perlpunk for suggesting this approach.
We change all similar places for consistency and safety, although as of
right now no other tests happen to have the same issue.

Signed-off-by: Adam Williamson <awilliam@redhat.com>