Return the error message the Downloader up the stack
We are currently swallowing the error message when anything during the download
& extraction process fails and simply report to the user that:
Downloading "$url" failed because of too many download errors

However, that is potentially completely wrong, as we also support extracting
files in downloader->download() and that process can fail due to a completely
different reason. The actual error message is then only visible in the log file
and needs admin access to the webui.

With this change we expose the error message in the webui as well, thereby
allowing the user to diagnose issues like this easier themselves without having
access to the logs.

This should also help diagnose issues like
https://progress.opensuse.org/issues/94997.