Avoid perl warning in string concatenation
Issue: https://progress.opensuse.org/issues/95730

We try to search for a Job with $self->id and no clone_id, and update
it with the new clone_id.
If no rows are affected then it means that $self had already set a
clone_id in the database, but we didn't reload the values of self, that's why
it is undef and producing a warning.

Instead we search for the Job again now.