Increase the waiting time for job results in scheduler-full test
We observed that t/05-scheduler-full.t fails more often than
it used too (it is in "unstable" already)

In one of the cases we wait for a job to be incompleted. We use
a active wait with a timeout, but seems that in environments with
a high load this timeout is not enough.

This PR increase the timeout from 20 seg. to 30 seg.

https://progress.opensuse.org/issues/95848