Remove unnecessary "SCALABILITY_TEST" variable
The test "43-scheduling-and-worker-scalability.t" is stable enough – at
least not the most unstable one ;) – and it does not use any special
ressources so we can remove the variable "SCALABILITY_TEST". This also
simplifies the setup for developers that want to call the test
explicitly as no special variable needs to be set anymore.