Fix OPENQA_CONFIG in OpenQA::UserAgent to override, not add config search path
Same is an other places if the environment variable OPENQA_CONFIG is
defined config files should be read from there and only there, not add
the path as an additional search path. This also allows to override with
an explicit empty configuration directory if desired, e.g. if one wants
to ensure that no private API credentials are read.