t: Fix 40-openqa-clone-job to not read personal config files
If there is a file ~/.config/openqa/client.conf matching
openqa.opensuse.org the test will fail because API key+secret are found
when not expected. As we can override the config directory with an empty
selection no valid config will be found in all cases which fixes the
test in such circumstances.