Fix "not last X" needle searches to include "never" results
If you set the "Limit Use:" or "Limit Match:" filter in the
Needles admin UI page to one of the "not last X" options, the
results don't include needles for which the "last use" or "last
match" value is "never" (shown as 'never' in the UI, represented
in the DB as null and in the abstraction layer as undef). This
is a bigger problem than it may seem at first glance because the
value can be 'never' for needles which *have* been seen/matched,
but only a long time ago and all those tests have now been
garbage-collected.

This fixes that, and also enhances the test to check these cases
(using both the 'custom' option and a pre-rolled option, as
those actually hit two different code paths). It also corrects
the existing tests a bit.

Signed-off-by: Adam Williamson <awilliam@redhat.com>