Add self-deduced git repo construction
As part of the changes for the investigation tab with links
to the commits, it was neccessity to find a way to make the
repo url self-agnostic, for the test repo as well as the needles.

This is what the purpose of `gitrepodir()` in *Utils.pm*. Running that,
it will return something like
"https://github.com/b10n1k/os-autoinst-distri-opensuse/commit/"
constructed by the of the local **.git/config** and the info under the `origin`
branch.
`origin` might not be the one that is set so this may need to be configurable,
but i leave it for later.

The reason of this approach is because:
- avoid static variables. In this case i could set the variables in the js and
  ask to use each one correspondingly.
- avoid openqa setup for every instance.
  Similar with the previous but the variables would have to adjust for every
  instance and finally passed to somehow to the js.

Also i stash the variables (*needlesgiturl* and *testgiturl*) inside `_show`.
i think this will give access to the variables also for other utulities.
However i come up with it because i strungled to stash
the variables from `investigation()`.

Signed-off-by: ybonatakis <ybonatakis@suse.com>