Optimize querying progress of running jobs by utilizing module columns
With OSD data (275 running jobs) the loading time for the particular AJAX
route goes down from ~ 500 ms to ~ 70 ms which is around 7 times faster.
It of course depends on the number of jobs and modules.

This simplification means there are some slight changes but those should
not matter in practice:

* The currently running module is completely ignored. This shouldn't matter
  much because we don't know how far the currently running module has
  progressed anyways. (For an example, see the change in `t/ui/01-list.t`.)
* If a test has no currently running module anymore it was previously
  shown as '100 %' in the short time until the job is set to uploading or
  done anyways. Now the ratio of processed to skipped modules is shown in
  that time.