Fix permission errors when running tidy with docker-compose setup
Otherwise one gets the following error messages when running tidy because
the database directory for the docker-compose setup is not accessible:

```
find: ‘./container/webui/workdir/db’: Keine Berechtigung
```

Due to the strict error handling of the script this prevents tidy to work.
This change simply prevents `find` from traversing the directory.