Improve logging for artefact uploads
It'll look like this:

```
[debug]   Uploading artefact bar
[warning] Error uploading bar: 500 response: Foo
[warning] Uploading artefact bar (attempts remaining: 4/5)
…
[warning] Uploading artefact bar (attempts remaining: 1/5)
[warning] All 5 upload attempts have failed for bar
[error]   Error uploading bar: 500 response: Foo
```

So the reason for failing attempts is logged and it is clear that a new
upload attempt starts. Previously the message

```
[warning] Upload attempts remaining: 4/5 for bar
```

was logged but not `Uploading artefact bar` anymore which made it look
like no further attempt was actually made.

See https://progress.opensuse.org/issues/94531