Trigger minion tasks when jobs are done
The new config section minion_task_triggers initially supports one type
of trigger which is on_job_done. The value is one or more minion task
names, such as limit_results_and_logs which is what we want in this
specific scenario. We also considered adding limit_assets in the
future, which is why this is more future-proof than a specific option
just for the current use case.
This cam be seen as an alternative to systemd timers which enqueue the
same tasks but at fixed intervals, although both can be used at the same
time. The aforementioned tasks already implement locks so it's safe to
enqueue them frequently. But this behavior depends on the specific task.

Fixes: https://progress.opensuse.org/issues/97763