docs: Treat database migration as advanced topic
This commit moves the database migration section to the advanced chapter
to keep the basic installation simpler for new users. Existing and
advanced users might need to follow a database version migration but
there are also use cases of fresh reinstalls if the test results in the
database are not worth to keep or container setups with externally
maintained databases.

This is a follow-up to commit 413abd4,
https://github.com/os-autoinst/openQA/pull/4171
which introduced the section about database migration.