Add Documentation to some systemd units
I'm sure there are more references that could be usefully added to these
(it's a space-separated list. see: systemd.unit(5)), and there are
several other units that could also use a Documentation setting where I
don't know enough about them, and the docs, to find a URL to use, but at
least this is a start.

Forwarded: no

Gbp-Pq: Name 0009-Add-Documentation-to-some-systemd-units.patch