Speed up asset cleanup by only computing asset sizes if not known yet
This means we won't track if an asset changes in size but I guess that's
not supposed to happen anyways. However, if a user deletes an asset
manually (not within the API or web UI) we would still assume it is there
until it would get deleted as part of the actual cleanup.