Speed up asset cleanup by only checking fixed status if not seen before
This means we need to do yet another check whether the asset is fixed
before deleting it to ensure we don't delete an asset which a user moved
to the "fixed" directory after openQA started tracking it. However, that
check should be quite cheap because at this point we're querying the
asset's location anyways.