templates: Add comma to prevent ambiguous state+timestamp combinations
Reading a string like

```
State: uploading started about an hour ago
```

can be confused by users to mean that the upload started about an hour
ago when "started" refers to the overall job. Adding a comma should
prevent that ambiguity or at least make it less likely to misunderstand.

To properly apply this change following
https://docs.mojolicious.org/Mojo/Template#SYNTAX
I needed to use backslash characters to escape newlines in combination
with trimming whitespace around tags with `<%= ... =%>`.