Fix computation of missing assets if there are no relevant assets
Apparently DBIx produces invalid SQL if the array for the "or" condition
is empty so we need to handle this case specifically to prevent an SQL
exception.

Without the `return [] unless @assets_query;` the extended test fails in
the same way as in https://progress.opensuse.org/issues/99327 so this
change should fix this issue.