tools/tidy: Disable aligning assignments
i.e. keep adding just single space instead.

The reason to disable aligning assignments is that with it one cannot
write code which style is correct. Having all the time whitespace issue
in PR is just ridiculous (and tools/tidy is slow even with
--only-changed to run it frequently).

Also there were too many whitespace changes, that just polluted git
history (it was needed to disable whitespace change in git diff/log
output to see only relevant changes) and thus slowed down test bug
investigation.

The downside of this change is that all os-autotest projects need to be
reformatted.

Implemented by adding --no-valign to perltidy as it looks to be the best
option [1].

Follow-up to os-autoinst change [2].

[1] https://stackoverflow.com/a/40438126/1564946
[2] https://github.com/os-autoinst/os-autoinst/pull/1787

Signed-off-by: Petr Vorel <pvorel@suse.cz>