Fix race-condition in `t/05-scheduler-full.t` which can lead to failures
If the worker process is fast enough to actually take (and incomplete) the
job before it is stopped the job is in a state which that does not work for
the subsequent test. This change simply makes sure the job's state is as
required regardless whether the job has been taken before or not.

This should fix the first problem mentioned in
https://progress.opensuse.org/issues/98898.