Prevent confusion about "priority"
For the openQA "priority" higher numbers mean actually "less priority"
so we considered different words to use, e.g. "niceness" as described on
https://en.wikipedia.org/wiki/Nice_%28Unix%29 but is also a very
Unix-specific concept. We might prefer more generic terms. Also such
rename would be a breaking change and would need to provide
backward-compatibility and/or a migration path.

Instead here I opted in documentation, help texts and comments to better
express that the specified value is *only* a value and not the priority
per se which is computed internally based on multiple factors.

This commit also adds a help text for the two -/+ buttons where
previously was no help text at all.

This commit was motivated by recurring user questions about the meaning
and the wish to clarify.